# 8345219: C2: x86_64 should not go to interpreter stubs for NaNs handling