@shipilev 2024-12-03
> This fix looks correct by looking on original changes.

Thank you! I need one more (R)eviewer :)