@eme64 2024-12-04
@shipilev Generally looks reasonable, though I'm not very familiar with this part of the code.

One question: Could we add an IR test that would show that no interpreter stubs are emitted, but instead whatever it is now emitting?