@shipilev 2024-12-04
> One question: Could we add an IR test that would show that no interpreter stubs are emitted, but instead whatever it is now emitting?

We could, but let's not do this in this PR. I already jumped the gun here a little with adding a benchmark to show the effect we have :)