@eme64 2024-12-04
Nice, that was quick!
Do we not need to restrict the IR rules? Or will that not fail on `IA32`?