@shipilev 2024-12-04
Seems to work on x86_64, AArch64, as well as on GHA. @eme64, are you good with this then?