@shipilev 2024-12-04
> I just launched testing, I don't think @vnkozlov did that already. Better to hold off until RDP1 / JDK25 fork anyway.

Thanks! I would like to have this fix in JDK 24, though: it is a (minor) performance optimization, and I would want to know if it breaks anything in ATR before we pull this block out completely when removing x86_32.